Refactor: change folder structure REST API V3 routes (#7876)
