+++ src/API/REST/V3/Routes/Campaigns/GetCampaignRevenue.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Give\Campaigns\Routes;
+namespace Give\API\REST\V3\Routes\Campaigns;
 
 use DateInterval;
 use DatePeriod;
 use DateTime;
 use DateTimeInterface;
 use Exception;
+use Give\API\REST\V3\Routes\Campaigns\ValueObjects\CampaignRoute;
 use Give\API\RestRoute;
 use Give\Campaigns\CampaignDonationQuery;
 use Give\Campaigns\Models\Campaign;
-use Give\Campaigns\ValueObjects\CampaignRoute;
 use WP_REST_Response;
 use WP_REST_Server;
