+++ src/API/REST/V3/Routes/Donors/ValueObjects/DonorAnonymousMode.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Give\Donations\ValueObjects;
+namespace Give\API\REST\V3\Routes\Donors\ValueObjects;
 
 use Give\Framework\Support\ValueObjects\Enum;
 
 /**
  * @since 4.0.0
  *
- * @method static DonationAnonymousMode EXCLUDE()
- * @method static DonationAnonymousMode INCLUDED()
- * @method static DonationAnonymousMode REDACTED()
+ * @method static DonorAnonymousMode EXCLUDE()
+ * @method static DonorAnonymousMode INCLUDED()
+ * @method static DonorAnonymousMode REDACTED()
  * @method bool isExcluded()
  * @method bool isIncluded()
  * @method bool isRedacted()
  */
-class DonationAnonymousMode extends Enum
+class DonorAnonymousMode extends Enum
 {
     const EXCLUDED = 'exclude';
     const INCLUDED = 'include';
