+++ src/Campaigns/ListTable/Routes/DeleteCampaignListTable.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Give\Campaigns\Routes;
+namespace Give\Campaigns\ListTable\Routes;
 
+use Give\API\REST\V3\Routes\Campaigns\ValueObjects\CampaignRoute;
 use Give\API\RestRoute;
 use Give\Campaigns\ListTable\CampaignsListTable;
 use Give\Campaigns\Repositories\CampaignRepository;
-use Give\Campaigns\ValueObjects\CampaignRoute;
 use Give\Framework\Exceptions\Primitives\Exception;
 use WP_Error;
 use WP_REST_Request;
