+++ tests/Unit/API/REST/V3/Routes/Donations/GetDonationsRouteTest.php
@@ -1,17 +1,16 @@
 <?php
 
-namespace Unit\Donors\Routes;
+namespace Unit\API\REST\V3\Routes\Donations;
 
 use Exception;
+use Give\API\REST\V3\Routes\Donations\RegisterDonationRoutes;
+use Give\API\REST\V3\Routes\Donations\ValueObjects\DonationRoute;
+use Give\API\REST\V3\Routes\Donors\ValueObjects\DonorRoute;
 use Give\Campaigns\Models\Campaign;
 use Give\Donations\Models\Donation;
 use Give\Donations\ValueObjects\DonationMetaKeys;
 use Give\Donations\ValueObjects\DonationMode;
 use Give\Donations\ValueObjects\DonationStatus;
-use Give\Donors\Models\Donor;
-use Give\Donors\Routes\RegisterDonorRoutes;
-use Give\Donors\ValueObjects\DonorRoute;
-use Give\Framework\Database\DB;
 use Give\Framework\Support\ValueObjects\Money;
 use Give\Helpers\Hooks;
 use Give\Tests\RestApiTestCase;
@@ -22,7 +21,7 @@ use WP_REST_Server;
 /**
  * @since 4.0.0
  */
-class GetDonorsRouteTest extends RestApiTestCase
+class GetDonationsRouteTest extends RestApiTestCase
 {
     use RefreshDatabase;
 
@@ -31,7 +30,7 @@ class GetDonorsRouteTest extends RestApiTestCase
      */
     public function setUp(): void
     {
-        Hooks::addAction('rest_api_init', RegisterDonorRoutes::class);
+        Hooks::addAction('rest_api_init', RegisterDonationRoutes::class);
 
         parent::setUp();
     }
@@ -39,26 +38,26 @@ class GetDonorsRouteTest extends RestApiTestCase
     /**
      * @since 4.0.0
      */
-    public function testGetDonorsShouldReturnAllModelProperties()
+    public function testGetDonationShouldReturnAllModelProperties()
     {
-        $newAdminUser = $this->factory()->user->create(
+        $donation = $this->createDonation1();
+
+        $newAdminUser = self::factory()->user->create(
             [
                 'role' => 'administrator',
-                'user_login' => 'testGetDonorsShouldReturnAllModelProperties',
-                'user_pass' => 'testGetDonorsShouldReturnAllModelProperties',
-                'user_email' => 'testGetDonorsShouldReturnAllModelProperties@test.com',
+                'user_login' => 'testGetDonationShouldReturnAllModelProperties',
+                'user_pass' => 'testGetDonationShouldReturnAllModelProperties',
+                'user_email' => 'testGetDonationShouldReturnAllModelProperties@test.com',
             ]
         );
+
         wp_set_current_user($newAdminUser);
 
-        /** @var  Donor $donor */
-        $donor = Donor::factory()->create();
-
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
+
         $request->set_query_params(
             [
-                'onlyWithDonations' => false,
                 'includeSensitiveData' => true,
             ]
         );
@@ -71,21 +70,38 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         // TODO: show shape of DateTime objects
         $createdAtJson = json_encode($data[0]['createdAt']);
+        $updatedAtJson = json_encode($data[0]['updatedAt']);
 
         $this->assertEquals(200, $status);
         $this->assertEquals([
-            'id' => $donor->id,
-            'userId' => $donor->userId,
+            'id' => $donation->id,
+            'campaignId' => $donation->campaignId,
+            'formId' => $donation->formId,
+            'formTitle' => $donation->formTitle,
+            'purchaseKey' => $donation->purchaseKey,
+            'donorIp' => $donation->donorIp,
             'createdAt' => json_decode($createdAtJson, true),
-            'name' => $donor->name,
-            'firstName' => $donor->firstName,
-            'lastName' => $donor->lastName,
-            'email' => $donor->email,
-            'phone' => $donor->phone,
-            'prefix' => $donor->prefix,
-            'additionalEmails' => $donor->additionalEmails,
-            'totalAmountDonated' => $donor->totalAmountDonated->toArray(),
-            'totalNumberOfDonations' => $donor->totalNumberOfDonations,
+            'updatedAt' => json_decode($updatedAtJson, true),
+            'status' => $donation->status->getValue(),
+            'type' => $donation->type->getValue(),
+            'mode' => $donation->mode->getValue(),
+            'amount' => $donation->amount->toArray(),
+            'feeAmountRecovered' => $donation->feeAmountRecovered ? $donation->feeAmountRecovered->toArray() : null,
+            'exchangeRate' => $donation->exchangeRate,
+            'gatewayId' => $donation->gatewayId,
+            'donorId' => $donation->donorId,
+            'honorific' => $donation->honorific,
+            'firstName' => $donation->firstName,
+            'lastName' => $donation->lastName,
+            'email' => $donation->email,
+            'phone' => $donation->phone,
+            'subscriptionId' => $donation->subscriptionId,
+            'billingAddress' => $donation->billingAddress ? $donation->billingAddress->toArray() : null,
+            'anonymous' => $donation->anonymous,
+            'levelId' => $donation->levelId,
+            'gatewayTransactionId' => $donation->gatewayTransactionId,
+            'company' => $donation->company,
+            'comment' => $donation->comment,
         ], $data[0]);
     }
 
@@ -94,17 +110,12 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldNotIncludeSensitiveData()
+    public function testGetDonationsShouldNotIncludeSensitiveData()
     {
-        Donor::factory()->create();
+        $this->createDonation1();
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
-        $request->set_query_params(
-            [
-                'onlyWithDonations' => false,
-            ]
-        );
 
         $response = $this->dispatchRequest($request);
 
@@ -112,10 +123,11 @@ class GetDonorsRouteTest extends RestApiTestCase
         $data = $response->get_data();
 
         $sensitiveProperties = [
-            'userId',
+            'donorIp',
             'email',
             'phone',
-            'additionalEmails',
+            'billingAddress',
+            'purchaseKey'
         ];
 
         $this->assertEquals(200, $status);
@@ -127,25 +139,25 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldIncludeSensitiveData()
+    public function testGetDonationsShouldIncludeSensitiveData()
     {
         $newAdminUser = $this->factory()->user->create(
             [
                 'role' => 'administrator',
-                'user_login' => 'testGetDonorsShouldIncludeSensitiveData',
-                'user_pass' => 'testGetDonorsShouldIncludeSensitiveData',
-                'user_email' => 'testGetDonorsShouldIncludeSensitiveData@test.com',
+                'user_login' => 'testGetDonationsShouldReturnSensitiveData',
+                'user_pass' => 'testGetDonationsShouldReturnSensitiveData',
+                'user_email' => 'testGetDonationsShouldReturnSensitiveData@test.com',
             ]
         );
         wp_set_current_user($newAdminUser);
 
-        Donor::factory()->create();
+        $this->createDonation1();
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
+
         $request->set_query_params(
             [
-                'onlyWithDonations' => false,
                 'includeSensitiveData' => true,
             ]
         );
@@ -156,10 +168,11 @@ class GetDonorsRouteTest extends RestApiTestCase
         $data = $response->get_data();
 
         $sensitiveProperties = [
-            'userId',
+            'donorIp',
             'email',
             'phone',
-            'additionalEmails',
+            'billingAddress',
+            'purchaseKey'
         ];
 
         $this->assertEquals(200, $status);
@@ -171,25 +184,25 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldReturn403ErrorWhenNotAdminUserIncludeSensitiveData()
+    public function testGetDonationsShouldReturn403ErrorWhenNotAdminUserIncludeSensitiveData()
     {
         $newSubscriberUser = $this->factory()->user->create(
             [
                 'role' => 'subscriber',
-                'user_login' => 'testGetDonorsShouldReturn403ErrorSensitiveData',
-                'user_pass' => 'testGetDonorsShouldReturn403ErrorSensitiveData',
-                'user_email' => 'testGetDonorsShouldReturn403ErrorSensitiveData@test.com',
+                'user_login' => 'testGetDonationShouldReturn403ErrorSensitiveData',
+                'user_pass' => 'testGetDonationShouldReturn403ErrorSensitiveData',
+                'user_email' => 'testGetDonationShouldReturn403ErrorSensitiveData@test.com',
             ]
         );
         wp_set_current_user($newSubscriberUser);
 
-        Donor::factory()->create();
+        $this->createDonation1();
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
+
         $request->set_query_params(
             [
-                'onlyWithDonations' => false,
                 'includeSensitiveData' => true,
             ]
         );
@@ -206,96 +219,23 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldReturnOnlyDonorsWithDonations()
+    public function testGetDonationsWithPagination()
     {
-        DB::query("DELETE FROM " . DB::prefix('give_donors'));
+        Donation::query()->delete();
 
-        /** @var Campaign $campaign */
-        $campaign = Campaign::factory()->create();
+        $donation1 = $this->createDonation1();
+        $donation2 = $this->createDonation2();
 
-        $donor1 = $this->createDonor1WithDonation($campaign->id);
-        $donor2 = Donor::factory()->create();
-
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
-        $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
-        $request->set_query_params(
-            [
-                'onlyWithDonations' => true,
-            ]
-        );
-
-        $response = $this->dispatchRequest($request);
-
-        $status = $response->get_status();
-        $data = $response->get_data();
-
-        $this->assertEquals(200, $status);
-        $this->assertEquals(1, count($data));
-        $this->assertEquals($donor1->id, $data[0]['id']);
-    }
-
-    /**
-     * @since 4.0.0
-     *
-     * @throws Exception
-     */
-    public function testGetDonorsShouldReturnDonorsWithOrWithoutDonations()
-    {
-        DB::query("DELETE FROM " . DB::prefix('give_donors'));
-
-        /** @var Campaign $campaign */
-        $campaign = Campaign::factory()->create();
-
-        $donor1 = $this->createDonor1WithDonation($campaign->id);
-        $donor2 = Donor::factory()->create();
-
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
-        $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
-        $request->set_query_params(
-            [
-                'onlyWithDonations' => false,
-                'direction' => 'ASC',
-            ]
-        );
-
-        $response = $this->dispatchRequest($request);
-
-        $status = $response->get_status();
-        $data = $response->get_data();
-
-        $this->assertEquals(200, $status);
-        $this->assertEquals(2, count($data));
-        $this->assertEquals($donor1->id, $data[0]['id']);
-        $this->assertEquals($donor2->id, $data[1]['id']);
-    }
-
-    /**
-     * @since 4.0.0
-     *
-     * @throws Exception
-     */
-    public function testGetDonorsWithPagination()
-    {
-        DB::query("DELETE FROM " . DB::prefix('give_donors'));
-
-        /** @var  Donor $donor1 */
-        $donor1 = Donor::factory()->create();
-
-        /** @var  Donor $donor2 */
-        $donor2 = Donor::factory()->create();
-
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
 
         $request->set_query_params(
             [
-                'onlyWithDonations' => false,
                 'page' => 1,
                 'per_page' => 1,
                 'direction' => 'ASC',
             ]
         );
-
         $response = $this->dispatchRequest($request);
 
         $status = $response->get_status();
@@ -304,12 +244,12 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(1, count($data));
-        $this->assertEquals($donor1->id, $data[0]['id']);
+        $this->assertEquals($donation1->id, $data[0]['id']);
         $this->assertEquals(2, $headers['X-WP-Total']);
         $this->assertEquals(2, $headers['X-WP-TotalPages']);
+
         $request->set_query_params(
             [
-                'onlyWithDonations' => false,
                 'page' => 2,
                 'per_page' => 1,
                 'direction' => 'ASC',
@@ -323,7 +263,7 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(1, count($data));
-        $this->assertEquals($donor2->id, $data[0]['id']);
+        $this->assertEquals($donation2->id, $data[0]['id']);
         $this->assertEquals(2, $headers['X-WP-Total']);
         $this->assertEquals(2, $headers['X-WP-TotalPages']);
     }
@@ -333,17 +273,15 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsByCampaignId()
+    public function testGetDonationsByCampaignId()
     {
-        Donation::query()->delete();
-
         /** @var Campaign $campaign */
         $campaign = Campaign::factory()->create();
 
-        $donor1 = $this->createDonor1WithDonation($campaign->id);
-        $donor2 = $this->createDonor2WithDonation($campaign->id);
+        $donation1 = $this->createDonation1($campaign->id);
+        $donation2 = $this->createDonation2($campaign->id);
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
         $request->set_query_params(
             [
@@ -358,8 +296,8 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(2, count($data));
-        $this->assertEquals($donor1->id, $data[0]['id']);
-        $this->assertEquals($donor2->id, $data[1]['id']);
+        $this->assertEquals($donation1->id, $data[0]['id']);
+        $this->assertEquals($donation2->id, $data[1]['id']);
     }
 
     /**
@@ -367,18 +305,18 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldNotIncludeAnonymousDonors()
+    public function testGetDonationsShouldNotIncludeAnonymousDonations()
     {
         Donation::query()->delete();
 
-        /** @var Campaign $campaign */
-        $campaign = Campaign::factory()->create();
+        $donation1 = $this->createDonation1();
 
-        $donor1 = $this->createDonor1WithDonation($campaign->id);
-        $donor2 = $this->createDonor2WithDonation($campaign->id, true);
+        // This anonymous donation should NOT be returned to the data array.
+        $donation2 = $this->createDonation2(0, true);
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
+
         $response = $this->dispatchRequest($request);
 
         $status = $response->get_status();
@@ -386,7 +324,7 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(1, count($data));
-        $this->assertEquals($donor1->id, $data[0]['id']);
+        $this->assertEquals($donation1->id, $data[0]['id']);
     }
 
     /**
@@ -394,31 +332,30 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldIncludeAnonymousDonors()
+    public function testGetDonationsShouldIncludeAnonymousDonations()
     {
         $newAdminUser = $this->factory()->user->create(
             [
                 'role' => 'administrator',
-                'user_login' => 'testGetDonorsShouldIncludeAnonymousDonors',
-                'user_pass' => 'testGetDonorsShouldIncludeAnonymousDonors',
-                'user_email' => 'testGetDonorsShouldIncludeAnonymousDonors@test.com',
+                'user_login' => 'testGetDonationsShouldIncludeAnonymousDonations',
+                'user_pass' => 'testGetDonationsShouldIncludeAnonymousDonations',
+                'user_email' => 'testGetDonationsShouldIncludeAnonymousDonations@test.com',
             ]
         );
         wp_set_current_user($newAdminUser);
 
         Donation::query()->delete();
 
-        /** @var Campaign $campaign */
-        $campaign = Campaign::factory()->create();
+        $donation1 = $this->createDonation1();
 
-        $donor1 = $this->createDonor1WithDonation($campaign->id);
-        $donor2 = $this->createDonor2WithDonation($campaign->id, true);
+        // This anonymous donation should be returned to the data array.
+        $donation2 = $this->createDonation2(0, true);
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
         $request->set_query_params(
             [
-                'anonymousDonors' => 'include',
+                'anonymousDonations' => 'include',
                 'direction' => 'ASC',
             ]
         );
@@ -430,8 +367,8 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(2, count($data));
-        $this->assertEquals($donor1->id, $data[0]['id']);
-        $this->assertEquals($donor2->id, $data[1]['id']);
+        $this->assertEquals($donation1->id, $data[0]['id']);
+        $this->assertEquals($donation2->id, $data[1]['id']);
     }
 
     /**
@@ -439,32 +376,27 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldReturn403ErrorWhenNotAdminUserIncludeAnonymousDonors()
+    public function testGetDonationsShouldReturn403ErrorWhenNotAdminUserIncludeAnonymousDonations()
     {
         $newSubscriberUser = $this->factory()->user->create(
             [
                 'role' => 'subscriber',
-                'user_login' => 'testGetDonorsShouldReturn403ErrorAnonymousDonors',
-                'user_pass' => 'testGetDonorsShouldReturn403ErrorAnonymousDonors',
-                'user_email' => 'testGetDonorsShouldReturn403ErrorAnonymousDonors@test.com',
+                'user_login' => 'testGetDonationShouldReturn403ErrorAnonymousDonation',
+                'user_pass' => 'testGetDonationShouldReturn403ErrorAnonymousDonation',
+                'user_email' => 'testGetDonationShouldReturn403ErrorAnonymousDonation@test.com',
             ]
         );
         wp_set_current_user($newSubscriberUser);
 
         Donation::query()->delete();
 
-        /** @var Campaign $campaign */
-        $campaign = Campaign::factory()->create();
+        $this->createDonation2(0, true);
 
-        $this->createDonor1WithDonation($campaign->id);
-        $this->createDonor2WithDonation($campaign->id, true);
-
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
         $request->set_query_params(
             [
-                'anonymousDonors' => 'include',
-                'direction' => 'ASC',
+                'anonymousDonations' => 'include',
             ]
         );
 
@@ -480,21 +412,20 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsShouldRedactAnonymousDonors()
+    public function testGetDonationsShouldRedactAnonymousDonations()
     {
         Donation::query()->delete();
 
-        /** @var Campaign $campaign */
-        $campaign = Campaign::factory()->create();
+        $donation1 = $this->createDonation1();
 
-        $donor1 = $this->createDonor1WithDonation($campaign->id);
-        $donor2 = $this->createDonor2WithDonation($campaign->id, true);
+        // This anonymous donation should be returned to the data array.
+        $donation2 = $this->createDonation2(0, true);
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $route = '/' . DonationRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
         $request->set_query_params(
             [
-                'anonymousDonors' => 'redact',
+                'anonymousDonations' => 'redact',
                 'direction' => 'ASC',
             ]
         );
@@ -506,15 +437,15 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(2, count($data));
-        $this->assertEquals($donor1->id, $data[0]['id']);
-        $this->assertEquals(0, $data[1]['id']);
+        $this->assertEquals($donation1->id, $data[0]['id']);
+        $this->assertEquals($donation2->id, $data[1]['id']);
 
         $anonymousDataRedacted = [
-            //'id', // This property is Checked above...
-            'name',
+            'donorId',
+            'honorific',
             'firstName',
             'lastName',
-            'prefix',
+            'company',
         ];
 
         foreach ($anonymousDataRedacted as $property) {
@@ -529,9 +460,9 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    public function testGetDonorsSortedByColumns($sortableColumn)
+    public function testGetDonationsSortedByColumns($sortableColumn)
     {
-        DB::query("DELETE FROM " . DB::prefix('give_donors'));
+        Donation::query()->delete();
 
         /** @var Campaign $campaign1 */
         $campaign1 = Campaign::factory()->create();
@@ -539,11 +470,12 @@ class GetDonorsRouteTest extends RestApiTestCase
         /** @var Campaign $campaign2 */
         $campaign2 = Campaign::factory()->create();
 
-        $donor1 = $this->createDonor1WithDonation($campaign1->id);
-        $donor2 = $this->createDonor2WithDonation($campaign1->id);
-        $donor3 = $this->createDonor3WithDonation($campaign2->id);
 
-        $route = '/' . DonorRoute::NAMESPACE . '/donors';
+        $donation1 = $this->createDonation1($campaign1->id);
+        $donation2 = $this->createDonation2($campaign1->id);
+        $donation3 = $this->createDonation3($campaign2->id);
+
+        $route = '/' . DonorRoute::NAMESPACE . '/donations';
         $request = new WP_REST_Request(WP_REST_Server::READABLE, $route);
 
         /**
@@ -565,9 +497,9 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(3, count($data));
-        $this->assertEquals($donor1->{$sortableColumn}, $data[0][$sortableColumn]);
-        $this->assertEquals($donor2->{$sortableColumn}, $data[1][$sortableColumn]);
-        $this->assertEquals($donor3->{$sortableColumn}, $data[2][$sortableColumn]);
+        $this->assertEquals($donation1->{$sortableColumn}, $data[0][$sortableColumn]);
+        $this->assertEquals($donation2->{$sortableColumn}, $data[1][$sortableColumn]);
+        $this->assertEquals($donation3->{$sortableColumn}, $data[2][$sortableColumn]);
 
         $request->set_query_params(
             [
@@ -586,8 +518,8 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(2, count($data));
-        $this->assertEquals($donor1->{$sortableColumn}, $data[0][$sortableColumn]);
-        $this->assertEquals($donor2->{$sortableColumn}, $data[1][$sortableColumn]);
+        $this->assertEquals($donation1->{$sortableColumn}, $data[0][$sortableColumn]);
+        $this->assertEquals($donation2->{$sortableColumn}, $data[1][$sortableColumn]);
 
         /**
          * Descendant Direction
@@ -608,9 +540,9 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(3, count($data));
-        $this->assertEquals($donor3->{$sortableColumn}, $data[0][$sortableColumn]);
-        $this->assertEquals($donor2->{$sortableColumn}, $data[1][$sortableColumn]);
-        $this->assertEquals($donor1->{$sortableColumn}, $data[2][$sortableColumn]);
+        $this->assertEquals($donation3->{$sortableColumn}, $data[0][$sortableColumn]);
+        $this->assertEquals($donation2->{$sortableColumn}, $data[1][$sortableColumn]);
+        $this->assertEquals($donation1->{$sortableColumn}, $data[2][$sortableColumn]);
 
         $request->set_query_params(
             [
@@ -629,8 +561,8 @@ class GetDonorsRouteTest extends RestApiTestCase
 
         $this->assertEquals(200, $status);
         $this->assertEquals(2, count($data));
-        $this->assertEquals($donor2->{$sortableColumn}, $data[0][$sortableColumn]);
-        $this->assertEquals($donor1->{$sortableColumn}, $data[1][$sortableColumn]);
+        $this->assertEquals($donation2->{$sortableColumn}, $data[0][$sortableColumn]);
+        $this->assertEquals($donation1->{$sortableColumn}, $data[1][$sortableColumn]);
     }
 
     /**
@@ -641,11 +573,13 @@ class GetDonorsRouteTest extends RestApiTestCase
         return [
             ['id'],
             ['createdAt'],
-            ['name'],
+            ['updatedAt'],
+            ['status'],
+            ['amount'],
+            ['feeAmountRecovered'],
+            ['donorId'],
             ['firstName'],
             ['lastName'],
-            ['totalAmountDonated'],
-            ['totalNumberOfDonations'],
         ];
     }
 
@@ -654,30 +588,24 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    private function createDonor1WithDonation(int $campaignId = 0, bool $anonymous = false): Donor
+    private function createDonation1(int $campaignId = 0, bool $anonymous = false): Donation
     {
         /** @var  Donation $donation1 */
         $donation1 = Donation::factory()->create([
             'status' => DonationStatus::COMPLETE(),
             'anonymous' => $anonymous,
+            'amount' => new Money(100, 'USD'),
+            'feeAmountRecovered' => new Money(10, 'USD'),
+            'firstName' => 'A',
+            'lastName' => 'A',
             'mode' => DonationMode::LIVE(),
         ]);
-        $donor1 = $donation1->donor;
-
-        $donor1->firstName = 'A';
-        $donor1->lastName = 'A';
-        $donor1->name = 'A A';
-        $donor1->totalAmountDonated = new Money(100, 'USD');
-        $donor1->totalNumberOfDonations = 1;
-        $donor1->save();
-
-        give()->payment_meta->update_meta($donation1->id, DonationMetaKeys::DONOR_ID, $donor1->id);
 
         if ($campaignId) {
             give()->payment_meta->update_meta($donation1->id, DonationMetaKeys::CAMPAIGN_ID, $campaignId);
         }
 
-        return Donor::find($donor1->id);
+        return $donation1;
     }
 
     /**
@@ -685,30 +613,24 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    private function createDonor2WithDonation(int $campaignId = 0, bool $anonymous = false): Donor
+    private function createDonation2(int $campaignId = 0, bool $anonymous = false): Donation
     {
         /** @var  Donation $donation2 */
         $donation2 = Donation::factory()->create([
             'status' => DonationStatus::COMPLETE(),
             'anonymous' => $anonymous,
+            'amount' => new Money(200, 'USD'),
+            'feeAmountRecovered' => new Money(20, 'USD'),
+            'firstName' => 'B',
+            'lastName' => 'B',
             'mode' => DonationMode::LIVE(),
         ]);
-        $donor2 = $donation2->donor;
-
-        $donor2->firstName = 'B';
-        $donor2->lastName = 'B';
-        $donor2->name = 'B B';
-        $donor2->totalAmountDonated = new Money(200, 'USD');
-        $donor2->totalNumberOfDonations = 2;
-        $donor2->save();
-
-        give()->payment_meta->update_meta($donation2->id, DonationMetaKeys::DONOR_ID, $donor2->id);
 
         if ($campaignId) {
             give()->payment_meta->update_meta($donation2->id, DonationMetaKeys::CAMPAIGN_ID, $campaignId);
         }
 
-        return Donor::find($donor2->id);
+        return $donation2;
     }
 
     /**
@@ -716,29 +638,23 @@ class GetDonorsRouteTest extends RestApiTestCase
      *
      * @throws Exception
      */
-    private function createDonor3WithDonation(int $campaignId = 0, bool $anonymous = false): Donor
+    private function createDonation3(int $campaignId = 0, bool $anonymous = false): Donation
     {
         /** @var  Donation $donation3 */
         $donation3 = Donation::factory()->create([
             'status' => DonationStatus::COMPLETE(),
             'anonymous' => $anonymous,
+            'amount' => new Money(300, 'USD'),
+            'feeAmountRecovered' => new Money(30, 'USD'),
+            'firstName' => 'C',
+            'lastName' => 'C',
             'mode' => DonationMode::LIVE(),
         ]);
-        $donor3 = $donation3->donor;
-
-        $donor3->firstName = 'C';
-        $donor3->lastName = 'C';
-        $donor3->name = 'C C';
-        $donor3->totalAmountDonated = new Money(300, 'USD');
-        $donor3->totalNumberOfDonations = 3;
-        $donor3->save();
-
-        give()->payment_meta->update_meta($donation3->id, DonationMetaKeys::DONOR_ID, $donor3->id);
 
         if ($campaignId) {
             give()->payment_meta->update_meta($donation3->id, DonationMetaKeys::CAMPAIGN_ID, $campaignId);
         }
 
-        return Donor::find($donor3->id);
+        return $donation3;
     }
 }
